requirements.txt
# Victory AI - Production Dependencies
# =====================================

# Core Data Science
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
joblib>=1.3.0

# FastAPI Backend (Production)
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
slowapi>=0.1.9

# API & Web Requests
requests>=2.31.0

# Environment Variables
python-dotenv>=1.0.0

# Optional: Auto-Training System
schedule>=1.1.0
